Rename 'components/DataViewTable.vue' to 'components/index/_shared/DataView/Table.vue'
