+++ components/index/CardsMonitoring/PositiveRate/Chart.vue
@@ -108,13 +108,13 @@ import Vue from 'vue'
 import { TranslateResult } from 'vue-i18n'
 import { ThisTypedComponentOptionsWithRecordProps } from 'vue/types/options'
 
-import DataViewTable, {
-  TableHeader,
-  TableItem,
-} from '@/components/DataViewTable.vue'
 import ScrollableChart from '@/components/index/_shared/charts/_base/ScrollableChart.vue'
 import DataView from '@/components/index/_shared/DataView.vue'
 import DataViewDataSetPanel from '@/components/index/_shared/DataView/DataSetPanel.vue'
+import DataViewTable, {
+  TableHeader,
+  TableItem,
+} from '@/components/index/_shared/DataView/Table.vue'
 import {
   DisplayData,
   yAxesBgPlugin,
