+++ components/index/CardsReference/Metro/Chart.vue
@@ -48,11 +48,11 @@ import Vue from 'vue'
 import { ThisTypedComponentOptionsWithRecordProps } from 'vue/types/options'
 
 import AppLink from '@/components/_shared/AppLink.vue'
+import DataView from '@/components/index/_shared/DataView.vue'
 import DataViewTable, {
   TableHeader,
   TableItem,
-} from '@/components/DataViewTable.vue'
-import DataView from '@/components/index/_shared/DataView.vue'
+} from '@/components/index/_shared/DataView/Table.vue'
 import { DisplayData } from '@/plugins/vue-chart'
 import { getGraphSeriesStyle } from '@/utils/colors'
