+++ components/index/_shared/charts/BarChart.vue
@@ -67,14 +67,14 @@ import Vue from 'vue'
 import { ThisTypedComponentOptionsWithRecordProps } from 'vue/types/options'
 
 import OpenDataLink from '@/components/_shared/OpenDataLink.vue'
-import DataViewTable, {
-  TableHeader,
-  TableItem,
-} from '@/components/DataViewTable.vue'
 import DataSelector from '@/components/index/_shared/charts/_base/DataSelector.vue'
 import ScrollableChart from '@/components/index/_shared/charts/_base/ScrollableChart.vue'
 import DataView from '@/components/index/_shared/DataView.vue'
 import DataViewBasicInfoPanel from '@/components/index/_shared/DataView/BasicInfoPanel.vue'
+import DataViewTable, {
+  TableHeader,
+  TableItem,
+} from '@/components/index/_shared/DataView/Table.vue'
 import { DisplayData, yAxesBgPlugin } from '@/plugins/vue-chart'
 import { getGraphSeriesStyle } from '@/utils/colors'
 import { getComplementedDate } from '@/utils/formatDate'
